requirements.txt
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.3.0
openai>=1.30.0
anthropic>=0.26.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
streamlit>=1.35.0
python-dotenv>=1.0.0
tabulate>=0.9.0
rich>=13.7.0
click>=8.1.0
pytest>=8.0.0
